% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate.R
\name{create_plate}
\alias{create_plate}
\title{Creates and ELISA plate template}
\usage{
create_plate(
  plate_type = 96,
  n_plates = 19,
  plate_names = NULL,
  file_type = c("xlsx", "csv"),
  file_name = NULL
)
}
\arguments{
\item{plate_type}{xxx}

\item{n_plates}{xxx}

\item{plate_names}{xxx}

\item{file_type}{xxx}

\item{file_name}{xxx}
}
\value{
xxx
}
\description{
Creates and ELISA plate template
}
\examples{
\dontrun{
create_plate()
}
}
