% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_regression}
\alias{plot_regression}
\title{Plot of linear and dose response regression.}
\usage{
plot_regression(
  file = NULL,
  data = NULL,
  group,
  dose,
  response,
  regression_model = c("linear", "dose_response"),
  dose_response_type = c("stimulation", "inhibition"),
  xLog = FALSE,
  facet = FALSE,
  ...
)
}
\arguments{
\item{file}{An excel or csv file path containing data to be fitted in a tidy format.}

\item{data}{A dataframe or tibble of data to be fitted.}

\item{group}{The column containing the different categories in the data or file.}

\item{dose}{The column in the data or file that will be used in plotting
the X axis. If it is log10 transformed `xLog` argument should be FALSE.}

\item{response}{The column in the data or file that will be used in plotting
the Y axis.}

\item{regression_model}{The regression model; either "linear" or "dose_response".
The default is "linear".}

\item{dose_response_type}{If `regression_model` is "dose_response" then specify
whether the response is stimulatory ("stimulation") or inhibitory ("inhibition").
The default is "stimulation".}

\item{xLog}{A logical value that determines whether the X axis will be log
transformed or not. The default is FALSE.}

\item{facet}{A logical value of whether to facet or not. The default is FALSE.}

\item{...}{Other arguments to be passed to ggplot2.}
}
\value{
A ggplot2 object.
}
\description{
Plot of linear and dose response regression.
}
