% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_regression.R
\name{stat_regression}
\alias{stat_regression}
\title{User friendly statistical output of linear and dose response regression.}
\usage{
stat_regression(
  file = NULL,
  data = NULL,
  group,
  dose,
  response,
  regression_model = c("linear", "dose_response"),
  dose_response_type = c("stimulation", "inhibition"),
  doseLog = TRUE
)
}
\arguments{
\item{file}{An excel or csv file path containing data to be fitted in a tidy format.}

\item{data}{A dataframe or tibble of data to be fitted.}

\item{group}{The column containing the different categories in the data or file.}

\item{dose}{The column in the data or file that will be used in the RHS.}

\item{response}{The column in the data or file that will be used in the LHS.}

\item{regression_model}{The regression model; either "linear" or "dose_response".
The default is "linear".}

\item{dose_response_type}{If `regression_model` is "dose_response" then specify
whether the response is stimulatory ("stimulation") or inhibitory ("inhibition").
The default is "stimulation".}

\item{doseLog}{A logical value of whether the `dose` of the input data is log10
transformed or not. If data is log10 transformed then the value should be
TRUE otherwise FALSE. The default is TRUE.}
}
\value{
A list of tibbles containing statistical parameters of the fitted data.
}
\description{
User friendly statistical output of linear and dose response regression.
}
