% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elisa.R
\name{plot_elisa}
\alias{plot_elisa}
\title{xxx}
\usage{
plot_elisa(file, group, x, plot_group, xlog = FALSE, errorbars = FALSE, ...)
}
\arguments{
\item{file}{xxx}

\item{group}{xxx}

\item{x}{xxx}

\item{plot_group}{xxx}

\item{xlog}{xxx}

\item{errorbars}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
xxx
}
